nake the no script image its own component
